# api/api/openapi.yaml
openapi: 3.0.3
info:
  title: EHR API
  description: |
    Headless Electronic Health Record API with dual interfaces:
    - FHIR R4-compliant endpoints under /fhir
    - Operational REST endpoints under /api/v1
  version: 0.1.0
  contact:
    name: EHR Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /fhir/Encounter:
    get:
      tags: [Encounter]
      summary: Search encounters (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - $ref: '#/components/parameters/FHIRStatus'
        - name: class
          in: query
          schema:
            type: string
          description: Encounter class (e.g. AMB, IMP, EMER)
        - $ref: '#/components/parameters/FHIRDate'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Encounter]
      summary: Create encounter (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Encounter'
      responses:
        '201':
          description: Encounter created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Encounter'
        '400':
          $ref: '#/components/responses/BadRequest'
  /fhir/Encounter/{id}:
    get:
      tags: [Encounter]
      summary: Read encounter by FHIR ID
      parameters:
        - $ref: '#/components/parameters/FHIRResourceId'
      responses:
        '200':
          description: Encounter resource
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Encounter'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Encounter]
      summary: Update encounter (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRResourceId'
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Encounter'
      responses:
        '200':
          description: Encounter updated
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Encounter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Encounter]
      summary: Delete encounter (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRResourceId'
      responses:
        '204':
          description: Encounter deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── FHIR Observation ───────────────────────────────────────────────
...
  /api/v1/encounters:
    get:
      tags: [Encounter]
      summary: List encounters
      parameters:
        - name: patient_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by patient ID
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Paginated list of encounters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Encounter]
      summary: Create encounter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncounterInput'
      responses:
        '201':
          description: Encounter created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/encounters/{id}:
    get:
      tags: [Encounter]
      summary: Get encounter by ID
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Encounter details
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Encounter]
      summary: Update encounter
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Encounter updated
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [Encounter]
      summary: Delete encounter
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Encounter deleted
  /api/v1/encounters/{id}/status:
    patch:
      tags: [Encounter]
      summary: Update encounter status
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [planned, arrived, triaged, in-progress, onleave, finished, cancelled]
      responses:
        '200':
          description: Status updated
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Observations (Operational) ─────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. In development mode, the server accepts any
        well-formed JWT. In production, tokens are validated against
        the configured JWKS endpoint.
  parameters:
    FHIRResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: FHIR logical resource ID
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal UUID
    FHIRPatient:
      name: patient
      in: query
      schema:
        type: string
      description: Patient reference (e.g. Patient/123)
    FHIRStatus:
      name: status
      in: query
      schema:
        type: string
      description: Resource status code
    FHIRDate:
      name: date
      in: query
      schema:
        type: string
      description: Date search (supports prefixes eq, ne, gt, lt, ge, le, sa, eb)
...
    Count:
      name: _count
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of results per page (FHIR)
    Offset:
      name: _offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Result offset for pagination (FHIR)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of results per page
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Result offset for pagination
  responses:
    FHIRBundle:
      description: FHIR Bundle (searchset)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/Bundle'
    BadRequest:
      description: Invalid request
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: missing or invalid token
...
    NotFound:
      description: Resource not found
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
    InternalError:
      description: Internal server error
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
  schemas:
...
    Bundle:
      type: object
      properties:
        resourceType:
          type: string
          enum: [Bundle]
        type:
          type: string
          enum: [searchset]
        total:
          type: integer
          description: Total number of matching resources
        link:
          type: array
          items:
            type: object
            properties:
              relation:
                type: string
                enum: [self, next, previous]
              url:
                type: string
                format: uri
        entry:
          type: array
          items:
            type: object
            properties:
              fullUrl:
                type: string
                format: uri
              resource:
                type: object
                description: The FHIR resource
              search:
                type: object
                properties:
                  mode:
                    type: string
                    enum: [match, include, outcome]
    OperationOutcome:
      type: object
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          items:
            type: object
            required: [severity, code]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
                description: Error code (not-found, invalid, etc.)
              diagnostics:
                type: string
                description: Human-readable diagnostic message
...
    Meta:
      type: object
      properties:
        versionId:
          type: string
        lastUpdated:
          type: string
          format: date-time
    Reference:
      type: object
      properties:
        reference:
          type: string
          description: Relative reference (e.g. Patient/123)
        display:
          type: string
    CodeableConcept:
      type: object
      properties:
        coding:
          type: array
          items:
            $ref: '#/components/schemas/Coding'
        text:
          type: string
    Coding:
      type: object
      properties:
        system:
          type: string
          format: uri
        code:
          type: string
        display:
          type: string
    Period:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
...
    Encounter:
      type: object
      required: [resourceType, status, class]
      properties:
        resourceType:
          type: string
          enum: [Encounter]
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        status:
          type: string
          enum: [planned, arrived, triaged, in-progress, onleave, finished, cancelled, entered-in-error]
        class:
          $ref: '#/components/schemas/Coding'
        type:
          type: array
          items:
            $ref: '#/components/schemas/CodeableConcept'
        subject:
          $ref: '#/components/schemas/Reference'
        participant:
          type: array
          items:
            type: object
            properties:
              type:
                type: array
                items:
                  $ref: '#/components/schemas/CodeableConcept'
              individual:
                $ref: '#/components/schemas/Reference'
        period:
          $ref: '#/components/schemas/Period'
        reasonCode:
          type: array
          items:
            $ref: '#/components/schemas/CodeableConcept'
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          description: Total number of matching records
        limit:
          type: integer
        offset:
          type: integer
...
    EncounterInput:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        patient_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [planned, arrived, triaged, in-progress, onleave, finished, cancelled]
        class_code:
          type: string
          description: Encounter class (AMB, IMP, EMER)
        type_code:
          type: string
        type_display:
          type: string
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        reason:
          type: string
...