# api/internal/domain/encounter/handler.go
package encounter

import (
	"net/http"

	"github.com/google/uuid"
	"github.com/labstack/echo/v4"

	"github.com/ehr/ehr/internal/platform/auth"
	"github.com/ehr/ehr/internal/platform/fhir"
	"github.com/ehr/ehr/pkg/pagination"
)

type Handler struct {
	svc *Service
}

func NewHandler(svc *Service) *Handler {
	return &Handler{svc: svc}
}

func (h *Handler) RegisterRoutes(api *echo.Group, fhirGroup *echo.Group) {
	// Read endpoints – admin, physician, nurse, registrar
	readGroup := api.Group("", auth.RequireRole("admin", "physician", "nurse", "registrar"))
	readGroup.GET("/encounters", h.ListEncounters)
	readGroup.GET("/encounters/:id", h.GetEncounter)
	readGroup.GET("/encounters/:id/participants", h.GetParticipants)
	readGroup.GET("/encounters/:id/diagnoses", h.GetDiagnoses)
	readGroup.GET("/encounters/:id/status-history", h.GetStatusHistory)

	// Write endpoints – admin, physician, nurse, registrar
	writeGroup := api.Group("", auth.RequireRole("admin", "physician", "nurse", "registrar"))
	writeGroup.POST("/encounters", h.CreateEncounter)
	writeGroup.PUT("/encounters/:id", h.UpdateEncounter)
	writeGroup.DELETE("/encounters/:id", h.DeleteEncounter)
	writeGroup.PATCH("/encounters/:id/status", h.UpdateEncounterStatus)
	writeGroup.POST("/encounters/:id/participants", h.AddParticipant)
	writeGroup.POST("/encounters/:id/diagnoses", h.AddDiagnosis)

	// FHIR read endpoints
	fhirRead := fhirGroup.Group("", auth.RequireRole("admin", "physician", "nurse", "registrar"))
	fhirRead.GET("/Encounter", h.SearchEncountersFHIR)
	fhirRead.GET("/Encounter/:id", h.GetEncounterFHIR)

	// FHIR write endpoints
	fhirWrite := fhirGroup.Group("", auth.RequireRole("admin", "physician", "nurse", "registrar"))
	fhirWrite.POST("/Encounter", h.CreateEncounterFHIR)
	fhirWrite.PUT("/Encounter/:id", h.UpdateEncounterFHIR)
	fhirWrite.DELETE("/Encounter/:id", h.DeleteEncounterFHIR)
}

// -- Operational Handlers --

func (h *Handler) CreateEncounter(c echo.Context) error {
	var enc Encounter
	if err := c.Bind(&enc); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	if err := h.svc.CreateEncounter(c.Request().Context(), &enc); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	return c.JSON(http.StatusCreated, enc)
}

func (h *Handler) GetEncounter(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	enc, err := h.svc.GetEncounter(c.Request().Context(), id)
	if err != nil {
		return echo.NewHTTPError(http.StatusNotFound, "encounter not found")
	}
	return c.JSON(http.StatusOK, enc)
}

func (h *Handler) ListEncounters(c echo.Context) error {
	pg := pagination.FromContext(c)

	if patientID := c.QueryParam("patient_id"); patientID != "" {
		pid, err := uuid.Parse(patientID)
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, "invalid patient_id")
		}
		encs, total, err := h.svc.ListEncountersByPatient(c.Request().Context(), pid, pg.Limit, pg.Offset)
		if err != nil {
			return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
		}
		return c.JSON(http.StatusOK, pagination.NewResponse(encs, total, pg.Limit, pg.Offset))
	}

	encs, total, err := h.svc.ListEncounters(c.Request().Context(), pg.Limit, pg.Offset)
	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	return c.JSON(http.StatusOK, pagination.NewResponse(encs, total, pg.Limit, pg.Offset))
}

func (h *Handler) UpdateEncounter(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	var enc Encounter
	if err := c.Bind(&enc); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	enc.ID = id
	if err := h.svc.UpdateEncounter(c.Request().Context(), &enc); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	return c.JSON(http.StatusOK, enc)
}

func (h *Handler) DeleteEncounter(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	if err := h.svc.DeleteEncounter(c.Request().Context(), id); err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	return c.NoContent(http.StatusNoContent)
}

func (h *Handler) UpdateEncounterStatus(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	var body struct {
		Status string `json:"status"`
	}
	if err := c.Bind(&body); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	if err := h.svc.UpdateEncounterStatus(c.Request().Context(), id, body.Status); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	return c.JSON(http.StatusOK, map[string]string{"status": body.Status})
}

func (h *Handler) AddParticipant(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	var p EncounterParticipant
	if err := c.Bind(&p); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	p.EncounterID = id
	if err := h.svc.AddParticipant(c.Request().Context(), &p); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	return c.JSON(http.StatusCreated, p)
}

func (h *Handler) GetParticipants(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	parts, err := h.svc.GetParticipants(c.Request().Context(), id)
	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	return c.JSON(http.StatusOK, parts)
}

func (h *Handler) AddDiagnosis(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	var d EncounterDiagnosis
	if err := c.Bind(&d); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	d.EncounterID = id
	if err := h.svc.AddDiagnosis(c.Request().Context(), &d); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
	}
	return c.JSON(http.StatusCreated, d)
}

func (h *Handler) GetDiagnoses(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	diags, err := h.svc.GetDiagnoses(c.Request().Context(), id)
	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	return c.JSON(http.StatusOK, diags)
}

func (h *Handler) GetStatusHistory(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}
	history, err := h.svc.GetStatusHistory(c.Request().Context(), id)
	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	return c.JSON(http.StatusOK, history)
}

// -- FHIR Encounter Handlers --

func (h *Handler) SearchEncountersFHIR(c echo.Context) error {
	pg := pagination.FromContext(c)
	params := map[string]string{}
	for _, key := range []string{"patient", "status", "class", "date"} {
		if v := c.QueryParam(key); v != "" {
			params[key] = v
		}
	}

	encs, total, err := h.svc.SearchEncounters(c.Request().Context(), params, pg.Limit, pg.Offset)
	if err != nil {
		return c.JSON(http.StatusInternalServerError, fhir.ErrorOutcome(err.Error()))
	}

	resources := make([]interface{}, len(encs))
	for i, e := range encs {
		resources[i] = e.ToFHIR()
	}
	return c.JSON(http.StatusOK, fhir.NewSearchBundle(resources, total, "/fhir/Encounter"))
}

func (h *Handler) GetEncounterFHIR(c echo.Context) error {
	enc, err := h.svc.GetEncounterByFHIRID(c.Request().Context(), c.Param("id"))
	if err != nil {
		return c.JSON(http.StatusNotFound, fhir.NotFoundOutcome("Encounter", c.Param("id")))
	}
	return c.JSON(http.StatusOK, enc.ToFHIR())
}

func (h *Handler) CreateEncounterFHIR(c echo.Context) error {
	var enc Encounter
	if err := c.Bind(&enc); err != nil {
		return c.JSON(http.StatusBadRequest, fhir.ErrorOutcome(err.Error()))
	}
	if err := h.svc.CreateEncounter(c.Request().Context(), &enc); err != nil {
		return c.JSON(http.StatusBadRequest, fhir.ErrorOutcome(err.Error()))
	}
	c.Response().Header().Set("Location", "/fhir/Encounter/"+enc.FHIRID)
	return c.JSON(http.StatusCreated, enc.ToFHIR())
}

func (h *Handler) UpdateEncounterFHIR(c echo.Context) error {
	existing, err := h.svc.GetEncounterByFHIRID(c.Request().Context(), c.Param("id"))
	if err != nil {
		return c.JSON(http.StatusNotFound, fhir.NotFoundOutcome("Encounter", c.Param("id")))
	}
	var enc Encounter
	if err := c.Bind(&enc); err != nil {
		return c.JSON(http.StatusBadRequest, fhir.ErrorOutcome(err.Error()))
	}
	enc.ID = existing.ID
	enc.FHIRID = existing.FHIRID
	enc.PatientID = existing.PatientID
	if err := h.svc.UpdateEncounter(c.Request().Context(), &enc); err != nil {
		return c.JSON(http.StatusBadRequest, fhir.ErrorOutcome(err.Error()))
	}
	return c.JSON(http.StatusOK, enc.ToFHIR())
}

func (h *Handler) DeleteEncounterFHIR(c echo.Context) error {
	enc, err := h.svc.GetEncounterByFHIRID(c.Request().Context(), c.Param("id"))
	if err != nil {
		return c.JSON(http.StatusNotFound, fhir.NotFoundOutcome("Encounter", c.Param("id")))
	}
	if err := h.svc.DeleteEncounter(c.Request().Context(), enc.ID); err != nil {
		return c.JSON(http.StatusInternalServerError, fhir.ErrorOutcome(err.Error()))
	}
	return c.NoContent(http.StatusNoContent)
}